requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
streamlit>=1.28.0
gspread>=5.12.0
google-auth>=2.23.0
